+++ idl/internal/error.go
@@ -20,32 +20,10 @@
 
 package internal
 
-import (
-	"bytes"
-	"fmt"
-)
+import "go.uber.org/thriftrw/ast"
 
-// parseError is an error type to keep track of any parse errors and the
-// positions they occur at.
-type parseError struct {
-	Errors map[int][]string
-}
-
-func newParseError() parseError {
-	return parseError{Errors: make(map[int][]string)}
-}
-
-func (pe parseError) add(line int, msg string) {
-	pe.Errors[line] = append(pe.Errors[line], msg)
-}
-
-func (pe parseError) Error() string {
-	var buffer bytes.Buffer
-	buffer.WriteString("parse error\n")
-	for line, msgs := range pe.Errors {
-		for _, msg := range msgs {
-			buffer.WriteString(fmt.Sprintf("  line %d: %s\n", line, msg))
-		}
-	}
-	return buffer.String()
+// ParseError holds a parse error and the position that caused it.
+type ParseError struct {
+	Pos ast.Position
+	Err error
 }